Fix test + s3_endpoint_alias usage
